requirements.txt
numpy 
tensorflow
torch
ultralytics
opencv-python
Pillow
onnxruntime
rembg
flask